+++ src/services/scanner/scanner.py
@@ -6,7 +6,7 @@ import asyncio
 from .strategies.light_cone_strategy import LightConeStrategy
 from .strategies.relic_strategy import RelicStrategy
 from pynput.keyboard import Key
-from utils.helpers import image_to_string, resource_path, preprocess_char_count_img
+from utils import image_to_string, resource_path, preprocess_char_count_img
 import pyautogui
 from .character_parser import CharacterParser
 from config.character_scan import CHARACTER_NAV_DATA
